Initial support of DSE for cstar_perf_tool
